Turn on caching & don't memoize dataloaders
I'm not sure what the purpose of memoizing the dataloaders was, but it requires
turning off caching, and it also caused a bug when passing arguments to
belongsToMany fields that made them only work the first time. Since caching also
enables deduplication, it seems like a good trade to turn off memoization to get
caching.